feature: 新增空间权限校验

+++ app/script/service/headers_service.py
@@ -6,8 +6,9 @@
 from app.common.decorators.service import http_service
 from app.common.decorators.transaction import transactional
 from app.common.id_generator import new_id
-from app.common.validator import check_not_exists
 from app.common.validator import check_exists
+from app.common.validator import check_not_exists
+from app.common.validator import check_workspace_permission
 from app.script.dao import http_header_dao as HttpHeaderDao
 from app.script.dao import http_header_template_dao as HttpHeaderTemplateDao
 from app.script.model import THttpHeader
@@ -19,7 +20,7 @@ log = get_logger(__name__)
 
 
 @http_service
-def query_http_header_template_list(req):
+def query_httpheader_template_list(req):
     # 条件分页查询
     pagination = HttpHeaderTemplateDao.select_list(
         workspaceNo=req.workspaceNo,
@@ -41,7 +42,7 @@ def query_http_header_template_list(req):
 
 
 @http_service
-def query_http_header_template_all(req):
+def query_httpheader_template_all(req):
     # 条件查询
     items = HttpHeaderTemplateDao.select_all(
         workspaceNo=req.workspaceNo,
@@ -62,10 +63,13 @@ def query_http_header_template_all(req):
 
 @http_service
 @transactional
-def create_http_header_template(req):
+def create_httpheader_template(req):
     # 查询模板
     template = HttpHeaderTemplateDao.select_by_workspace_and_name(req.workspaceNo, req.templateName)
     check_not_exists(template, '模板已存在')
+
+    # 校验空间权限
+    check_workspace_permission(req.workspaceNo)
 
     # 新增模板
     template_no = new_id()
@@ -81,10 +85,13 @@ def create_http_header_template(req):
 
 @http_service
 @transactional
-def modify_http_header_template(req):
+def modify_httpheader_template(req):
     # 查询模板
     template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
     check_exists(template, '模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
 
     # 更新模板
     template.update(
@@ -95,11 +102,12 @@ def modify_http_header_template(req):
 
 @http_service
 @transactional
-def remove_http_header_template(req):
+def remove_httpheader_template(req):
     # 查询模板
     template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
     check_exists(template, '模板不存在')
-
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
     # 删除模板下的所有请求头
     HttpHeaderDao.delete_all_by_template(req.templateNo)
     # 删除模板
@@ -116,6 +124,9 @@ def create_http_header(req):
     # 查询模板
     template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
     check_exists(template, '模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
 
     # 新增请求头
     header_no = new_id()
@@ -138,6 +149,13 @@ def modify_http_header(req):
     header = HttpHeaderDao.select_by_no(req.headerNo)
     check_exists(header, '请求头不存在')
 
+    # 查询模板
+    template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
+    check_exists(template, '模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
+
     # 更新请求头
     header.update(
         HEADER_NAME=req.headerName,
@@ -153,6 +171,13 @@ def remove_http_header(req):
     header = HttpHeaderDao.select_by_no(req.headerNo)
     check_exists(header, '请求头不存在')
 
+    # 查询模板
+    template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
+    check_exists(template, '模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
+
     # 删除请求头
     header.delete()
 
@@ -163,6 +188,13 @@ def enable_http_header(req):
     # 查询请求头
     header = HttpHeaderDao.select_by_no(req.headerNo)
     check_exists(header, '请求头不存在')
+
+    # 查询模板
+    template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
+    check_exists(template, '模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
 
     # 启用请求头
     header.update(
@@ -176,6 +208,13 @@ def disable_http_header(req):
     # 查询请求头
     header = HttpHeaderDao.select_by_no(req.headerNo)
     check_exists(header, '请求头不存在')
+
+    # 查询模板
+    template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
+    check_exists(template, '模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
 
     # 禁用请求头
     header.update(
@@ -231,6 +270,9 @@ def create_http_headers(req):
     template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
     check_exists(template, '模板不存在')
 
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
+
     for header in req.headerList:
         # 跳过请求头为空的数据
         if not header.headerName:
@@ -254,6 +296,12 @@ def create_http_headers(req):
 @http_service
 @transactional
 def modify_http_headers(req):
+    # 查询模板
+    template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
+    check_exists(template, '模板不存在')
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
+    #  遍历更新请求头
     for header in req.headerList:
         # 跳过请求头为空的数据
         if not header.headerName:
@@ -287,23 +335,31 @@ def modify_http_headers(req):
 @http_service
 @transactional
 def remove_http_headers(req):
+    # 查询模板
+    template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
+    check_exists(template, '模板不存在')
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
     # 批量删除请求头
-    HttpHeaderDao.delete_in_no(req.list)
+    HttpHeaderDao.delete_in_no(req.headerNumberedList)
 
 
 @http_service
 @transactional
-def duplicate_http_header_template(req):
+def duplicate_httpheader_template(req):
     # 查询请求头模板
     template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
     check_exists(template, '请求头模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
 
     # 复制请求头模板
     template_no = new_id()
     THttpHeaderTemplate.insert(
         WORKSPACE_NO=template.WORKSPACE_NO,
         TEMPLATE_NO=template_no,
-        TEMPLATE_NAME=template.TEMPLATE_NAME + ' copy',
+        TEMPLATE_NAME=f'{template.TEMPLATE_NAME} copy',
         TEMPLATE_DESC=template.TEMPLATE_DESC
     )
 
@@ -324,17 +380,20 @@ def duplicate_http_header_template(req):
 
 @http_service
 @transactional
-def copy_http_header_template_to_workspace(req):
+def copy_httpheader_template_to_workspace(req):
     # 查询请求头模板
     template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
     check_exists(template, '请求头模板不存在')
+
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
 
     # 复制请求头模板
     template_no = new_id()
     THttpHeaderTemplate.insert(
         WORKSPACE_NO=req.workspaceNo,
         TEMPLATE_NO=template_no,
-        TEMPLATE_NAME=template.TEMPLATE_NAME + ' copy',
+        TEMPLATE_NAME=f'{template.TEMPLATE_NAME} copy',
         TEMPLATE_DESC=template.TEMPLATE_DESC
     )
 
@@ -355,9 +414,11 @@ def copy_http_header_template_to_workspace(req):
 
 @http_service
 @transactional
-def move_http_header_template_to_workspace(req):
+def move_httpheader_template_to_workspace(req):
     # 查询请求头模板
     template = HttpHeaderTemplateDao.select_by_no(req.templateNo)
     check_exists(template, '请求头模板不存在')
-
+    # 校验空间权限
+    check_workspace_permission(template.WORKSPACE_NO)
+    # 移动请求头模板
     template.update(WORKSPACE_NO=req.workspaceNo)
