+++ app/script/service/variables_service.py
@@ -6,8 +6,9 @@
 from app.common.decorators.service import http_service
 from app.common.decorators.transaction import transactional
 from app.common.id_generator import new_id
-from app.common.validator import check_not_exists
 from app.common.validator import check_exists
+from app.common.validator import check_not_exists
+from app.common.validator import check_workspace_permission
 from app.script.dao import variable_dao as VariableDao
 from app.script.dao import variable_dataset_dao as VariableDatasetDao
 from app.script.enum import VariableDatasetType
@@ -70,16 +71,19 @@ def query_variable_dataset_all(req):
 @transactional
 def create_variable_dataset(req):
     # 查询变量集信息
-    varset = VariableDatasetDao.select_first(
+    dataset = VariableDatasetDao.select_first(
         WORKSPACE_NO=req.workspaceNo,
         DATASET_NAME=req.datasetName,
         DATASET_TYPE=req.datasetType
     )
-    check_not_exists(varset, '变量集已存在')
+    check_not_exists(dataset, '变量集已存在')
 
     # 变量集为ENVIRONMENT或CUSTOM时，工作空间编号不能为空
     if req.datasetType != VariableDatasetType.GLOBAL.value:
         check_exists(req.workspaceNo, '工作空间编号不能为空')
+
+    # 校验空间权限
+    check_workspace_permission(req.workspaceNo)
 
     # 新增变量集
     dataset_no = new_id()
@@ -99,11 +103,14 @@ def create_variable_dataset(req):
 @transactional
 def modify_variable_dataset(req):
     # 查询变量集信息
-    varset = VariableDatasetDao.select_by_no(req.datasetNo)
-    check_exists(varset, '变量集不存在')
+    dataset = VariableDatasetDao.select_by_no(req.datasetNo)
+    check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
 
     # 更新变量集信息
-    varset.update(
+    dataset.update(
         DATASET_NAME=req.datasetName,
         DATASET_DESC=req.datasetDesc
     )
@@ -115,7 +122,8 @@ def remove_variable_dataset(req):
     # 查询变量集信息
     dataset = VariableDatasetDao.select_by_no(req.datasetNo)
     check_exists(dataset, '变量集不存在')
-
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
     # 删除变量集下的所有变量
     VariableDao.delete_all_by_dataset(req.datasetNo)
     # 删除变量集
@@ -132,6 +140,9 @@ def create_variable(req):
     # 查询变量集信息
     dataset = VariableDatasetDao.select_by_no(req.datasetNo)
     check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
 
     # 新增变量
     var_no = new_id()
@@ -155,6 +166,13 @@ def modify_variable(req):
     variable = VariableDao.select_by_no(req.varNo)
     check_exists(variable, '变量不存在')
 
+    # 查询变量集信息
+    dataset = VariableDatasetDao.select_by_no(variable.DATASET_NO)
+    check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
+
     # 更新变量信息
     variable.update(
         VAR_NAME=req.varName,
@@ -171,6 +189,13 @@ def remove_variable(req):
     variable = VariableDao.select_by_no(req.varNo)
     check_exists(variable, '变量不存在')
 
+    # 查询变量集信息
+    dataset = VariableDatasetDao.select_by_no(variable.DATASET_NO)
+    check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
+
     # 删除变量
     variable.delete()
 
@@ -181,6 +206,13 @@ def enable_variable(req):
     # 查询变量信息
     variable = VariableDao.select_by_no(req.varNo)
     check_exists(variable, '变量不存在')
+
+    # 查询变量集信息
+    dataset = VariableDatasetDao.select_by_no(variable.DATASET_NO)
+    check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
 
     # 启用变量
     variable.update(
@@ -195,6 +227,13 @@ def disable_variable(req):
     variable = VariableDao.select_by_no(req.varNo)
     check_exists(variable, '变量不存在')
 
+    # 查询变量集信息
+    dataset = VariableDatasetDao.select_by_no(variable.DATASET_NO)
+    check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
+
     # 禁用变量
     variable.update(
         ENABLED=False
@@ -207,6 +246,13 @@ def update_current_value(req):
     # 查询变量信息
     variable = VariableDao.select_by_no(req.varNo)
     check_exists(variable, '变量不存在')
+
+    # 查询变量集信息
+    dataset = VariableDatasetDao.select_by_no(variable.DATASET_NO)
+    check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
 
     # 更新当前值
     variable.update(
@@ -264,6 +310,9 @@ def create_variables(req):
     dataset = VariableDatasetDao.select_by_no(req.datasetNo)
     check_exists(dataset, '变量集不存在')
 
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
+
     for vari in req.variableList:
         # 跳过变量名为空的数据
         if not vari.varName:
@@ -288,6 +337,12 @@ def create_variables(req):
 @http_service
 @transactional
 def modify_variables(req):
+    # 查询变量集信息
+    dataset = VariableDatasetDao.select_by_no(req.datasetNo)
+    check_exists(dataset, '变量集不存在')
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
+    # 遍历更新变量
     for vari in req.variableList:
         # 跳过变量名为空的数据
         if not vari.varName:
@@ -323,8 +378,13 @@ def modify_variables(req):
 @http_service
 @transactional
 def remove_variables(req):
+    # 查询变量集信息
+    dataset = VariableDatasetDao.select_by_no(req.datasetNo)
+    check_exists(dataset, '变量集不存在')
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
     # 批量删除变量
-    VariableDao.delete_in_no(req.list)
+    VariableDao.delete_in_no(req.variableNumberedList)
 
 
 @http_service
@@ -333,6 +393,9 @@ def duplicate_variable_dataset(req):
     # 查询变量集
     dataset = VariableDatasetDao.select_by_no(req.datasetNo)
     check_exists(dataset, '变量集不存在')
+
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
 
     # 复制变量集
     dataset_no = new_id()
@@ -368,6 +431,9 @@ def copy_variable_dataset_to_workspace(req):
     dataset = VariableDatasetDao.select_by_no(req.datasetNo)
     check_exists(dataset, '变量集不存在')
 
+    # 校验空间权限
+    check_workspace_permission(dataset.WORKSPACE_NO)
+
     # 复制变量集
     dataset_no = new_id()
     TVariableDataset.insert(
@@ -401,5 +467,7 @@ def move_variable_dataset_to_workspace(req):
     # 查询变量集
     dataset = VariableDatasetDao.select_by_no(req.datasetNo)
     check_exists(dataset, '变量集不存在')
-
+    # 校验空间权限
+    check_workspace_permission(req.workspaceNo)
+    # 移动变量集
     dataset.update(WORKSPACE_NO=req.workspaceNo)
